+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix `IdToken::getClaim` method to return default value if claim is not found
+
 ## [0.2.4] - 2026-02-19
 
 ### Removed