+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix `IdToken::getClaim` method to return default value if claim is not found
+- Fix `nonce` handling in `AuthCodeGrant` to only set if present in request
 
 ## [0.2.4] - 2026-02-19