+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.5] - 2026-04-03
+
 ### Fixed
 
 - Fix `IdToken::getClaim` method to return default value if claim is not found